Create helper.py
This helper file consists the graph plotting code. Graph plotting was done to visualize the improvements made by our model. To plot the graph Matplotlib was used which is a Python library to visualize the data. On the X-axis the number of games played by AI were plotted and on the Y-axis the scores were plotted.